feat: Add successful and failed API delegations to ConfirmationPage
The code changes in this commit add the functionality to display successful and failed API delegations on the ConfirmationPage. This is achieved by filtering the data received from the API and storing the successful and failed delegations separately. The color of the page is also updated based on the presence of failed or successful delegations.